Extract useClampedOverflow hook from ArticlePreview
The ResizeObserver + measurement state was 20 lines of plumbing in
ArticlePreview, which is already a monster. Extracted to a generic
hook in src/hooks/ so the consumer just declares "tell me if this
ref's content overflows" and the noise lives elsewhere — reusable for
any clamped element (titles, descriptions, anywhere we add a clamp
later).
